app/requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-multipart

# Vector Database
chromadb==0.4.18

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Embeddings and LLM
sentence-transformers==2.7.0
transformers==4.36.0
torch>=2.0.0
huggingface-hub>=0.20.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
